Home

QMTK

Welcome to QMTK (Quantum Many-body Toolkit)

The QMTK library is developed for physics system involves quantum many-body system, including condensed matter physics, quantum computing, quantum information and etc.

In most numerical and theoretical works, programs related to quantum physics and quantum many-body system could have a lot similar functionality, methods, and objects. We aims to provide physicist with a powerful toolkit in Julia language as its frontend interface.

QMTK.nn.Linear — Type.

Linear{T, WeightType, BiasType, InputType} <: AbstractBlock

Linear layer.

$z^{l+1} = W^l * z^{l} + b^l$

expression of gradients:

$\delta^{l} = (W^l)^T\delta^{l+1}$ $\delta^{l}_w = \delta^{l+1}(z^l)^T$ $\delta^{l}_b = \delta^{l+1}$

source