Basic Types

Basic Types

Basis

Basis is a useful object when describing a Hilbert space. In quantum physics, there are many kind of basis in different problems, including qubits, lattice site, infinity basis, etc.

The basis we mention here and implemented in this package is the tagged part. If you need to use Dirac ket operations, you can simply do it by hands with dot function in Julia's linear algebra stdlib. To be specific, the basis here means what is inside you ket expression.

$|basis\rangle$ or $\langle basis|$

Label

QMTK offers three kind of labels: Bit, Spin, Half for binary configurations.

QMTK.Bit — Type.

Bit <: AbstractBitSite

Binary State, Bit. Has two state: 0, 1

source

QMTK.Spin — Type.

Spin <: AbstractBitSite

Binary State, Spin. Has two state: -1, 1

source

QMTK.Half — Type.

Half <: AbstractBitSite

Binary State, Half. Has two state: -0.5, 0.5

source

Sites

The abstract hierachy for site is defined as any subtype of AbstractArray with label.

QMTK.AbstractSites — Type.

AbstractSites{Label, T, N} <: AbstractArray{T, N}

Sites are arrays with certain labels

source

Sites is a Julia native array with certain label that tells the program which kind of site it is.

QMTK.Sites — Type.

Sites{L <: SiteLabel, T, N} <: AbstractSites{L, T, N}

Lattice Sites are Julia Arrays with certain Label, it is able to use array interface.

TODO: type promote to normal arrays when mixed with normal arrays.

source

Sites is implemented to support most of Julia array interface, but since Julia do not have inherit of interface, not all operations is overloaded, please use method data, when you need to access its content in calculatoin.

QMTK.data — Function.

data(basis)

get data of the basis

source

SubSites

For small sites, it usually contains only a few configuration, and when we try to use a subset of Sites, it would be a waste to create a new Sites, SubSites offer a way to do this efficiently by telling the compiler how large it is in compile time (by Tuple).

Or to be simple, you can just interpret SubSites to be a Tuple with certain SiteLabel.